Add initial multi-literal indexing for integer induction clauses
