+++ Indexing/CodeTree.cpp
@@ -42,14 +42,6 @@ using namespace Kernel;
 
 //////////////// general datastructures ////////////////////
 
-CodeTree::LitInfo::LitInfo(Clause* cl, unsigned litIndex)
-: litIndex(litIndex), opposite(false)
-{
-  CALL("CodeTree::LitInfo::LitInfo");
-
-  ft=FlatTerm::create((*cl)[litIndex]);
-}
-
 void CodeTree::LitInfo::dispose()
 {
   ft->destroy();
